+++ src/nvim/message.c
@@ -149,7 +149,7 @@ int msg_attr(char_u *s, int attr) FUNC_ATTR_NONNULL_ARG(1)
   return msg_attr_keep(s, attr, FALSE);
 }
 
-int 
+int
 msg_attr_keep (
     char_u *s,
     int attr,
@@ -645,7 +645,7 @@ char_u *msg_may_trunc(int force, char_u *s)
   return s;
 }
 
-static void 
+static void
 add_msg_hist (
     char_u *s,
     int len,                        /* -1 for undetermined length */
@@ -1206,7 +1206,7 @@ void msg_make(char_u *arg)
  * This function is used to show mappings, where we want to see how to type
  * the character/string -- webb
  */
-int 
+int
 msg_outtrans_special (
     char_u *strstart,
     int from               /* TRUE for lhs of a mapping */
@@ -1796,7 +1796,7 @@ static int do_clear_sb_text = FALSE;    /* clear text on next msg */
 /*
  * Store part of a printed message for displaying when scrolling back.
  */
-static void 
+static void
 store_sb_text (
     char_u **sb_str,           /* start of string */
     char_u *s,                 /* just after string */
@@ -2154,7 +2154,7 @@ static int do_more_prompt(int typed_char)
           msg_scroll_up();
           inc_msg_scrolled();
           screen_fill((int)Rows - 2, (int)Rows - 1, 0,
-              (int)Columns, ' ', ' ', 0);
+              Columns, ' ', ' ', 0);
           mp_last = disp_sb_line((int)Rows - 2, mp_last);
           --toscroll;
         }
@@ -2163,7 +2163,7 @@ static int do_more_prompt(int typed_char)
       if (toscroll <= 0) {
         /* displayed the requested text, more prompt again */
         screen_fill((int)Rows - 1, (int)Rows, 0,
-            (int)Columns, ' ', ' ', 0);
+            Columns, ' ', ' ', 0);
         msg_moremsg(FALSE);
         continue;
       }
@@ -2363,11 +2363,11 @@ void msg_clr_eos_force(void)
   } else {
     if (cmdmsg_rl) {
       screen_fill(msg_row, msg_row + 1, 0, msg_col + 1, ' ', ' ', 0);
-      screen_fill(msg_row + 1, (int)Rows, 0, (int)Columns, ' ', ' ', 0);
+      screen_fill(msg_row + 1, (int)Rows, 0, Columns, ' ', ' ', 0);
     } else {
-      screen_fill(msg_row, msg_row + 1, msg_col, (int)Columns,
+      screen_fill(msg_row, msg_row + 1, msg_col, Columns,
           ' ', ' ', 0);
-      screen_fill(msg_row + 1, (int)Rows, 0, (int)Columns, ' ', ' ', 0);
+      screen_fill(msg_row + 1, (int)Rows, 0, Columns, ' ', ' ', 0);
     }
   }
 }
@@ -2623,7 +2623,7 @@ void msg_advance(int col)
  * A '&' in a button name becomes a shortcut, so each '&' should be before a
  * different letter.
  */
-int 
+int
 do_dialog (
     int type,
     char_u *title,
@@ -2715,7 +2715,7 @@ do_dialog (
  * Copy one character from "*from" to "*to", taking care of multi-byte
  * characters.  Return the length of the character in bytes.
  */
-static int 
+static int
 copy_char (
     char_u *from,
     char_u *to,
