+++ src/nvim/option.c
@@ -5544,7 +5544,7 @@ set_num_option (
     if (updating_screen) {
       *pp = old_value;
     } else if (full_screen) {
-      screen_resize((int)Columns, (int)Rows);
+      screen_resize(Columns, (int)Rows);
     } else {
       /* Postpone the resizing; check the size and cmdline position for
        * messages. */
@@ -6074,7 +6074,7 @@ showoptions (
                 && Columns + GAP >= INT_MIN + 3
              && (Columns + GAP - 3) / INC >= INT_MIN
              && (Columns + GAP - 3) / INC <= INT_MAX);
-      cols = (int)((Columns + GAP - 3) / INC);
+      cols = ((Columns + GAP - 3) / INC);
       if (cols == 0)
         cols = 1;
       rows = (item_count + cols - 1) / cols;
@@ -6474,11 +6474,11 @@ void comp_col(void)
   assert(sc_col >= 0
          && INT_MIN + sc_col <= Columns
          && Columns - sc_col <= INT_MAX);
-  sc_col = (int)(Columns - sc_col);
+  sc_col = (Columns - sc_col);
   assert(ru_col >= 0
          && INT_MIN + ru_col <= Columns
          && Columns - ru_col <= INT_MAX);
-  ru_col = (int)(Columns - ru_col);
+  ru_col = (Columns - ru_col);
   if (sc_col <= 0)              /* screen too narrow, will become a mess */
     sc_col = 1;
   if (ru_col <= 0)
